Move moveType into move.h and pre-shifted values
